+++ src/main/java/com/verdantartifice/primalmagick/common/tiles/devices/HoneyExtractorTileEntity.java
@@ -98,11 +98,11 @@ public class HoneyExtractorTileEntity extends TileInventoryPM implements MenuPro
     }
 
     @Override
-    public CompoundTag save(CompoundTag compound) {
+    protected void saveAdditional(CompoundTag compound) {
+        super.saveAdditional(compound);
         compound.putInt("SpinTime", this.spinTime);
         compound.putInt("SpinTimeTotal", this.spinTimeTotal);
         compound.put("ManaStorage", this.manaStorage.serializeNBT());
-        return super.save(compound);
     }
 
     @Override